requirements.txt
# Core ML Libraries - Un conjunto probado para compatibilidad
transformers==4.39.3
accelerate==0.29.3
bitsandbytes==0.43.0
peft==0.10.0
datasets==2.18.0

# La imagen base ya tiene torch, no lo necesitamos aquí.

# RAG y Scripts
sentence-transformers
faiss-cpu
numpy<2.0 # Importante para compatibilidad con librerías más antiguas
pyyaml
python-dotenv

# Logging
tensorboard
tensorboardX

# Dependencias secundarias
huggingface-hub

# --- Dependencias del Coder ---
fastapi
uvicorn[standard]
torch

# --- Dependencias del Orquestador ---
flask
requests
